+++ R/best_step_fit_bic.R
@@ -1,17 +1,17 @@
 #'
-#' best.step.fit.bic
+#' best_step_fit_bic
 #'
 #' finds the ML-best step model for one row and
 #' compares the posteriors of the found best step
 #' and non-step uniform model.
 #'
-#' See [friends.test] documentation for details.
+#' See [friends_test] documentation for details.
 #'
-#' @inheritParams step.fit.ln.likelihoods
+#' @inheritParams step_fit_ln_likelihoods
 #' @param prior.to.have.friends The prior for a row is important enough to
 #' have friendly columns
 #' @return a list of four values: \cr
-#' \code{step.models} is the value return by [step.fit.ln.likelihoods]
+#' \code{step.models} is the value return by [step_fit_ln_likelihoods]
 #' call the function start with
 #' \code{best.step.rank} is the rank value that makes the best step;
 #' it is not obligatory one on the \code{ranks} value.\cr
@@ -27,11 +27,11 @@
 #' all columns are listed in \code{columns.on.right} and
 #' \code{columns.on.left} is empty.
 #' @examples
-#' example(row.int.ranks)
-#' step <- best.step.fit.bic(TF.ranks[42, ], genes.no, 0.5)
-#' nostep <- best.step.fit.bic(TF.ranks[42, ], genes.no, 1E-50)
+#' example(row_int_ranks)
+#' step <- best_step_fit_bic(TF.ranks[42, ], genes.no, 0.5)
+#' nostep <- best_step_fit_bic(TF.ranks[42, ], genes.no, 1E-50)
 #' @export
-best.step.fit.bic <- function(ranks, max.possible.rank, prior.to.have.friends) {
+best_step_fit_bic <- function(ranks, max.possible.rank, prior.to.have.friends) {
     step.models <- .step_fit_compact(ranks, max.possible.rank)
     k <- length(ranks)